pyproject.toml
[tool.poetry]
name = "t-module_menu"
version = "0.1.0"
description = ""
authors = ["Tryd0g0lik <work80@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
flask = {extras = ["async"], version = "^3.0.3"}
django = ">=4.2.14"

[tool.poetry.group.dev.dependencies]
asyncio = "^3.4.3"
autohooks = "^24.2.0"
flake8 = "^7.1.1"
Django = "4.2.14"
django-webpack-loader = "3.1.0"
postgres = "4.0"
psycopg2-binary = "2.9.9"
djangorestframework = "3.15.2"
django-cors-headers = "4.4.0"
pre-commit = "^3.8.0"
django-bootstrap4 = "^24.4"
markdown = "^3.7"
pylint = "^3.3.1"
isort = "^5.13.2"
black = "^24.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
